Local transmission of B.1.1.7 in Austria
Press release by the ministry of health confirming sequencing of 46 cases of B.1.1.7 lineage in Austria on the 18th of January 2021. 
Including cases in a care home facility in Vienna. 
https://www.ots.at/presseaussendung/OTS_20210118_OTS0147/anschober-ergebnisse-der-vollsequenzierungen-bestaetigen-b117-verdacht-bei-grossteil-der-geprueften-faelle

Further confirmation of prevalence of B.1.1.7 in a sample of 66 out of 539 samples in Vienna on the 22nd of January:
https://www.wienerzeitung.at/nachrichten/chronik/wien/2089738-B.1.1.7-in-zwoelf-Prozent-der-Wiener-Proben.html